requirements.txt
Flask==2.2.5
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
transformers==4.35.2
torch==2.2.0
apscheduler==3.10.1
pandas==2.2.1
sqlalchemy==2.1.0
python-dotenv==1.0.0
vaderSentiment==3.3.3
